pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ytuploader"
version = "0.1.0"
description = "Windows 11 desktop app for editing OBS recordings and uploading to YouTube."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyQt6>=6.6",
  "google-api-python-client>=2.120.0",
  "google-auth>=2.28.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
]

[project.optional-dependencies]
dev = [
  "pyinstaller>=6.0.0",
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
